Refactor/dashboard (#8)
* feat: add slider const for dashboard

* reafactor: split star rating function into new util

* reafactor: split navigation list from dashboard  into new component

* reafactor: split image slide from dashboard  into new component

* reafactor: move dashboardsection component into components/dashboard

* feat: create flash sale timer component for flash sale section in dashboard

* refactor: split flash sale section from dashboard into new component

* refactor: cleaner code in dashboard using separated component
