+++ src/contants/slider.ts
@@ -1,0 +1,7 @@
+export const SLIDER = [
+    { label: 'iPhone 16 Pro Max', img: 'hero1.png', title: 'iPhone 16 Series', desc: 'Up to 10% off Voucher' },
+    { label: 'Bleu de Chanel', img: 'hero2.png', title: 'Bleu de Chanel', desc: 'Up to 20% off Voucher' },
+    { label: 'Vellaro Shoes', img: 'hero3.jpg', title: 'Vellaro Shoes', desc: 'Up to 30% off Voucher' },
+    { label: 'Orient Watch', img: 'hero4.jpg', title: 'Orient Watch', desc: 'Up to 40% off Voucher' },
+    { label: 'MF Serum', img: 'hero5.jpg', title: 'MF Serum', desc: 'Up to 50% off Voucher' },
+];
